# src/docs/swagger.yaml
definitions:
...
  github_com_aminazadbakht1_golang-clean-web-api_api_dto.CarModelColorResponse:
    properties:
      color:
        $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.ColorResponse'
      id:
        type: integer
    type: object
  github_com_aminazadbakht1_golang-clean-web-api_api_dto.CarModelCommentResponse:
    properties:
      carModelId:
        type: integer
      id:
        type: integer
      message:
        type: string
      user:
        $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.UserResponse'
    type: object
...
  github_com_aminazadbakht1_golang-clean-web-api_api_dto.ColorResponse:
    properties:
      hexCode:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  github_com_aminazadbakht1_golang-clean-web-api_api_dto.CreateCarModelColorRequest:
    properties:
      carModelId:
        type: integer
      colorId:
        type: integer
    required:
    - carModelId
    - colorId
    type: object
  github_com_aminazadbakht1_golang-clean-web-api_api_dto.CreateCarModelCommentRequest:
    properties:
      carModelId:
        type: integer
      message:
        maxLength: 100
        type: string
      userId:
        type: integer
    required:
    - carModelId
    - message
    type: object
...
  github_com_aminazadbakht1_golang-clean-web-api_api_dto.Filter:
    properties:
      filterType:
        description: text number
        type: string
      from:
        type: string
      to:
        type: string
      type:
        description: contains notContains equals notEqual startsWith lessThan lessThanOrEqual
          greaterThan greaterThanOrEqual inRange endsWith
        type: string
    type: object
...
  ? github_com_aminazadbakht1_golang-clean-web-api_api_dto.PagedList-github_com_aminazadbakht1_golang-clean-web-api_api_dto_CarModelColorResponse
  : properties:
      hasNextPage:
        type: boolean
      hasPreviousPage:
        type: boolean
      items:
        items:
          $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.CarModelColorResponse'
        type: array
      pageNumber:
        type: integer
      totalPages:
        type: integer
      totalRows:
        type: integer
    type: object
  ? github_com_aminazadbakht1_golang-clean-web-api_api_dto.PagedList-github_com_aminazadbakht1_golang-clean-web-api_api_dto_CarModelCommentResponse
  : properties:
      hasNextPage:
        type: boolean
      hasPreviousPage:
        type: boolean
      items:
        items:
          $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.CarModelCommentResponse'
        type: array
      pageNumber:
        type: integer
      totalPages:
        type: integer
      totalRows:
        type: integer
    type: object
...
  github_com_aminazadbakht1_golang-clean-web-api_api_dto.PaginationInputWithFilter:
    properties:
      filter:
        additionalProperties:
          $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.Filter'
        type: object
      pageNumber:
        type: integer
      pageSize:
        type: integer
      sort:
        items:
          $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.Sort'
        type: array
    type: object
...
  github_com_aminazadbakht1_golang-clean-web-api_api_dto.Sort:
    properties:
      colId:
        type: string
      sort:
        type: string
    type: object
  github_com_aminazadbakht1_golang-clean-web-api_api_dto.UpdateCarModelColorRequest:
    properties:
      carModelId:
        type: integer
      colorId:
        type: integer
    type: object
  github_com_aminazadbakht1_golang-clean-web-api_api_dto.UpdateCarModelCommentRequest:
    properties:
      message:
        maxLength: 100
        type: string
    required:
    - message
    type: object
...
  github_com_aminazadbakht1_golang-clean-web-api_api_dto.UserResponse:
    properties:
      email:
        type: string
      firstName:
        type: string
      id:
        type: integer
      lastName:
        type: string
      username:
        type: string
    type: object
  github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse:
    properties:
      error: {}
      result: {}
      resultCode:
        $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.ResultCode'
      success:
        type: boolean
      validationErrors:
        items:
          $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_validation.ValidationError'
        type: array
    type: object
  github_com_aminazadbakht1_golang-clean-web-api_api_helper.ResultCode:
    enum:
    - 0
    - 40001
    - 40101
    - 40301
    - 40401
    - 42901
    - 42902
    - 50001
    - 50002
    type: integer
    x-enum-varnames:
    - Success
    - ValidationError
    - AuthError
    - ForbiddenError
    - NotFoundError
    - LimiterError
    - OtpLimiterError
    - CustomRecovery
    - InternalError
  github_com_aminazadbakht1_golang-clean-web-api_api_validation.ValidationError:
    properties:
      message:
        type: string
      property:
        type: string
      tag:
        type: string
      value:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /v1/car-model-colors/:
    post:
      consumes:
      - application/json
      description: Create a CarModelColor
      parameters:
      - description: Create a CarModelColor
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.CreateCarModelColorRequest'
      responses:
        "201":
          description: CarModelColor response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.CarModelColorResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Create a CarModelColor
      tags:
      - CarModelColors
  /v1/car-model-colors/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a CarModelColor
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: response
          schema:
            $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Delete a CarModelColor
      tags:
      - CarModelColors
    get:
      consumes:
      - application/json
      description: Get a CarModelColor
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: CarModelColor response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.CarModelColorResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get a CarModelColor
      tags:
      - CarModelColors
    put:
      consumes:
      - application/json
      description: Update a CarModelColor
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      - description: Update a CarModelColor
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.UpdateCarModelColorRequest'
      responses:
        "200":
          description: CarModelColor response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.CarModelColorResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Update a CarModelColor
      tags:
      - CarModelColors
  /v1/car-model-colors/get-by-filter:
    post:
      consumes:
      - application/json
      description: Get CarModelColors
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.PaginationInputWithFilter'
      responses:
        "200":
          description: CarModelColor response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.PagedList-github_com_aminazadbakht1_golang-clean-web-api_api_dto_CarModelColorResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get CarModelColors
      tags:
      - CarModelColors
  /v1/car-model-comments/:
    post:
      consumes:
      - application/json
      description: Create a CarModelComment
      parameters:
      - description: Create a CarModelComment
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.CreateCarModelCommentRequest'
      responses:
        "201":
          description: CarModelComment response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.CarModelCommentResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Create a CarModelComment
      tags:
      - CarModelComments
  /v1/car-model-comments/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a CarModelComment
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: response
          schema:
            $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Delete a CarModelComment
      tags:
      - CarModelComments
    get:
      consumes:
      - application/json
      description: Get a CarModelComment
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: CarModelComment response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.CarModelCommentResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get a CarModelComment
      tags:
      - CarModelComments
    put:
      consumes:
      - application/json
      description: Update a CarModelComment
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      - description: Update a CarModelComment
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.UpdateCarModelCommentRequest'
      responses:
        "200":
          description: CarModelComment response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.CarModelCommentResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Update a CarModelComment
      tags:
      - CarModelComments
  /v1/car-model-comments/get-by-filter:
    post:
      consumes:
      - application/json
      description: Get CarModelComments
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.PaginationInputWithFilter'
      responses:
        "200":
          description: CarModelComment response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_dto.PagedList-github_com_aminazadbakht1_golang-clean-web-api_api_dto_CarModelCommentResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_aminazadbakht1_golang-clean-web-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get CarModelComments
      tags:
      - CarModelComments
...
securityDefinitions:
  AuthBearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"